Moved models to their own folder and general bug fixes
