+++ app/Orchid/Layouts/FarmerReport/FarmerReportEditLayout.php
@@ -2,9 +2,9 @@
 
 namespace App\Orchid\Layouts\FarmerReport;
 
-use App\Models\Crop;
+use App\Models\Crop\Crop;
 use App\Models\Farmland\Farmland;
-use App\Models\SeedStage;
+use App\Models\Crop\SeedStage;
 use App\Models\User;
 use Orchid\Screen\Field;
 use Orchid\Screen\Fields\Input;
