+++ app/Orchid/Screens/Farmland/FarmlandEditScreen.php
@@ -3,7 +3,7 @@
 namespace App\Orchid\Screens\Farmland;
 
 use App\Actions\Farmland\DeleteFarmland;
-use App\FarmerReport\Farmland\CreateFarmland;
+use App\Actions\Farmland\CreateFarmland;
 use App\Models\Farmland\Farmland;
 use App\Orchid\Layouts\Farmland\FarmlandEditFarmLayout;
 use App\Orchid\Layouts\Farmland\FarmlandEditMemberLayout;
